# specs/001-homelab-dashboard/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: HomeDash API
  version: 0.1.0
  description: |
    API contracts for HomeDash.

    Notes:
    - Cookie-based sessions are used for authentication.
    - CSRF protection is required for state-changing requests.
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/first-run/admin:
    post:
      tags: [auth]
      summary: Create the initial admin user (first run only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FirstRunAdminCreateRequest'
      responses:
        '201':
          description: Created and authenticated
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMe'
        '409':
          description: Users already exist (first-run disabled)
  /api/auth/login:
    post:
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMe'
        '401':
          description: Invalid credentials
  /api/auth/logout:
    post:
      tags: [auth]
      summary: Logout
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204':
          description: Logged out
  /api/auth/me:
    get:
      tags: [auth]
      summary: Current authenticated user
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMe'
        '401':
          description: Not authenticated
...
  /api/admin/icons/refresh:
    post:
      tags: [admin]
      summary: Admin-controlled icon refresh (best-effort fetch + cache)
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '202':
          description: Refresh started

  # ── Docker (043-docker-remote-auth) ─────────────────────────────────────────
  # No route here accepts an endpoint from the caller: it is resolved
  # server-side from the widget's stored connection. The former
  # POST /api/docker/ping is deliberately absent — it was unauthenticated.
  /api/docker/hosts:
    get:
      operationId: listDockerWidgetHosts
      tags: [docker]
      summary: List the ordered Docker hosts linked to a widget
      security:
        - cookieAuth: []
      parameters:
        - name: widgetInstanceId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ordered host identifiers and display names
        '401':
          description: Not authenticated
        '422':
          description: Malformed query
  /api/docker/containers:
    get:
      operationId: listDockerContainers
      tags: [docker]
      summary: List containers for one of a widget's linked Docker connections
      security:
        - cookieAuth: []
      parameters:
        - name: widgetInstanceId
          in: query
          required: true
          description: Identifies the widget that owns the selected connection.
          schema:
            type: string
        - name: connectionId
          in: query
          required: true
          description: Opaque Docker connection ID linked to the widget.
          schema:
            type: string
        - name: all
          in: query
          required: false
          description: Include non-running containers. Defaults to true.
          schema:
            type: string
            enum: [  'true', 'false' ]
      responses:
        '200':
          description: Container list
        '400':
          description: The stored endpoint is not a valid Docker endpoint
        '401':
          description: Not authenticated
        '409':
          description: The widget has no Docker connection configured
        '502':
          description: The endpoint is configured but the daemon did not answer
  /api/docker/action:
    post:
      operationId: dockerContainerAction
      tags: [docker]
      summary: Start, stop or restart a container
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [widgetInstanceId, connectionId, containerId, action]
              properties:
                widgetInstanceId:
                  type: string
                connectionId:
                  type: string
                containerId:
                  type: string
                  maxLength: 64
                action:
                  type: string
                  enum: [start, stop, restart]
      responses:
        '200':
          description: Action accepted
        '401':
          description: Not authenticated
        '403':
          description: Not an admin, or CSRF token missing or invalid
        '409':
          description: The widget has no Docker connection configured
        '422':
          description: Malformed body
        '502':
          description: The endpoint is configured but the daemon did not answer
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: homedash_session
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
  schemas:
...
    FirstRunAdminCreateRequest:
      type: object
      required: [username, displayName, password]
      properties:
        username:
          type: string
        displayName:
          type: string
        password:
          type: string
          format: password
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    AuthMe:
      type: object
      required: [user]
      properties:
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      required: [id, username, displayName, role]
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [admin, standard]
...